Use the library 'query-string' to serialize query params.
Don't need 'SearchResult' so move its code to 'Search' and remove it.

Use async/await to call the search api in the test.

Add more JSDoc.
