+++ oeq-ts-rest-api/package.json
@@ -30,6 +30,7 @@
   "dependencies": {
     "axios": "^0.19.2",
     "axios-cookiejar-support": "^1.0.0",
+    "query-string": "^6.12.1",
     "tough-cookie": "^4.0.0",
     "typescript-is": "^0.16.0"
   },
